+++ R/lb-improved.R
@@ -129,23 +129,10 @@ lb_improved <- function(x, y, window.size = NULL, norm = "L1", lower.env = NULL,
 # Loop without using native 'proxy' looping (to avoid multiple calculations of the envelope)
 # ========================================================================================================
 
-lb_improved_loop <- function(x, y = NULL, ...) {
+lb_improved_loop <- function(x, y = NULL, window.size = NULL, error.check = TRUE,
+                             force.symmetry = FALSE, norm = "L1", ...) {
 
-     ARGS <- list(...)
-     window.size <- ARGS$window.size
-     error.check <- ARGS$error.check
-     force.symmetry <- ARGS$force.symmetry
-     norm <- ARGS$norm
-
-     if (is.null(error.check))
-          error.check <- TRUE
-     if (is.null(force.symmetry))
-          force.symmetry <- FALSE
-
-     if (is.null(norm))
-          norm <- "L1"
-     else
-          norm <- match.arg(norm, c("L1", "L2"))
+     norm <- match.arg(norm, c("L1", "L2"))
 
      if (error.check)
           window.size <- consistency_check(window.size, "window")