+++ R/zscore.R
@@ -1,7 +1,7 @@
 #' Wrapper for z-normalization
 #'
 #' Wrapper for function \code{\link[base]{scale}} that returns zeros instead of \code{NaN}. It also
-#' supports a list of vectors and matrices.
+#' supports a list of vectors and a matrix input.
 #'
 #' @param x Data to normalize. Either a vector, a matrix where each row is to be normalized, or a list of vectors.
 #' @param ... Further arguments to pass to \code{\link[base]{scale}}.
@@ -10,6 +10,7 @@
 #' @return Normalized data.
 #'
 #' @export
+#'
 
 zscore <- function(x, ..., na.rm = FALSE) {